requirements.txt
langchain
elevenlabs
python-dotenv
streamlit
openai
